reorganized code for new structure
