+++ src/common.py
@@ -1,0 +1,6 @@
+import base64
+
+
+def decode_base64url(s: str) -> bytes:
+    pad = "=" * (-len(s) % 4)
+    return base64.urlsafe_b64decode(s + pad)
