+++ src/server/exceptions.py
@@ -1,0 +1,10 @@
+class ServerError(Exception):
+    pass
+
+
+class UrlParseError(ServerError):
+    pass
+
+
+class UnsupportedProtocolError(ServerError):
+    pass
